pyproject.toml
[project]
name = "backend-core"
version = "0.1.0"
description = "Backend Core Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "structlog>=25.5.0",
    "sentry-sdk>=2.47.0",
    "beanie>=2.0.1",
    "python-dotenv>=1.2.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "motor>=3.6.0",
    "email-validator>=2.0.0",
    "argon2-cffi>=25.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
strict = true
exclude = ["build"]
disallow_subclassing_any = false
disallow_untyped_decorators = false
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "tests.*"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
    "types-requests>=2.32.4.20250913",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=app --cov-report=term-missing"
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function call in argument defaults
]
